moving the navigator directory inside the Facet directory
https://coveord.atlassian.net/browse/JSUI-2367
