+++ src/ui/Facet/FacetSearchDropdownNavigation/SearchDropdownNavigatorFactory.ts
@@ -1,8 +1,8 @@
 import { ISearchDropdownNavigator, ISearchDropdownConfig, DefaultSearchDropdownNavigator } from './DefaultSearchDropdownNavigator';
 import { FacetSearchDropdownNavigator } from './FacetSearchDropdownNavigator';
-import { IFacetSearch } from '../Facet/IFacetSearch';
-import { CategoryFacet } from '../CategoryFacet/CategoryFacet';
-import { Facet } from '../Facet/Facet';
+import { IFacetSearch } from '../IFacetSearch';
+import { CategoryFacet } from '../../CategoryFacet/CategoryFacet';
+import { Facet } from '../Facet';
 
 export function SearchDropdownNavigatorFactory(facetSearch: IFacetSearch, config: ISearchDropdownConfig): ISearchDropdownNavigator {
   switch (facetSearch.facetType) {
